requirements.txt
asgiref==3.3.4
Django==3.2.3
django-js-asset==1.2.2
django-tinymce4-lite==1.8.0
jsmin==2.2.2
pytz==2021.1
sqlparse==0.4.1
